feat(shared-video/native): created ShareVideoConfirmDialog and unified actions
