+++ react/features/shared-video/components/index.native.ts
@@ -1,3 +1,4 @@
 // @ts-ignore
 export { default as SharedVideoDialog } from './native/SharedVideoDialog';
 export { default as SharedVideoButton } from './native/SharedVideoButton';
+export { default as ShareVideoConfirmDialog } from './native/ShareVideoConfirmDialog';
