+++ react/features/shared-video/components/index.web.ts
@@ -1,2 +1,3 @@
 export { default as SharedVideoDialog } from './web/SharedVideoDialog';
 export { default as SharedVideoButton } from './web/SharedVideoButton';
+export { default as ShareVideoConfirmDialog } from './web/ShareVideoConfirmDialog';
